Mixing runtime and python_codegen_support.py
