+++ src/schema_salad/metaschema.py
@@ -4,53 +4,833 @@
 # subject to the license of the original schema.
 from __future__ import annotations
 
-import copy
 import os
 import sys
 import uuid as _uuid__
-from collections.abc import Collection, MutableMapping, MutableSequence
-from typing import Any, ClassVar, Final, cast
-
+from collections.abc import Collection
 from mypy_extensions import trait
-from ruamel.yaml.comments import CommentedMap
+from typing import ClassVar
 
-from schema_salad.exceptions import ValidationException
 from schema_salad.runtime import (
-    LoadingOptions,
-    Saveable,
-    _AnyLoader,
-    _ArrayLoader,
-    _EnumLoader,
-    _ExpressionLoader,
-    _IdMapLoader,
-    _MapLoader,
-    _PrimitiveLoader,
-    _RecordLoader,
-    _SecondaryDSLLoader,
-    _TypeDSLLoader,
-    _UnionLoader,
-    _URILoader,
-    convert_typing,
-    expand_url,
-    extract_type,
     file_uri,
-    load_field,
     parse_errors,
     prefix_url,
     save,
     save_relative_uri,
-    _document_load,
 )
-from schema_salad.sourceline import SourceLine, add_lc_filename
-from schema_salad.utils import yaml_no_ts
 
 if sys.version_info >= (3, 11):
     from typing import Self
 else:
     from typing_extensions import Self
 
+import copy
+
+from collections.abc import MutableSequence, Sequence, MutableMapping
+from io import StringIO
+from itertools import chain
+from typing import Any, Final, cast, Generic, TypeVar
+from urllib.parse import urldefrag, urlsplit, urlunsplit
+
+from ruamel.yaml.comments import CommentedMap
+
+from schema_salad.exceptions import ValidationException, SchemaSaladException
+from schema_salad.runtime import LoadingOptions, convert_typing, extract_type, Saveable
+from schema_salad.sourceline import SourceLine, add_lc_filename
+from schema_salad.utils import yaml_no_ts  # requires schema-salad v8.2+
+
+S = TypeVar("S", bound=Saveable)
+
+
 _vocab: Final[dict[str, str]] = {}
 _rvocab: Final[dict[str, str]] = {}
+
+
+class _Loader:
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any | None:
+        pass
+
+
+class _AnyLoader(_Loader):
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        if doc is not None:
+            return doc
+        raise ValidationException("Expected non-null")
+
+
+class _PrimitiveLoader(_Loader):
+    def __init__(self, tp: type | tuple[type[str], type[str]]) -> None:
+        self.tp: Final = tp
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        if not isinstance(doc, self.tp):
+            raise ValidationException(f"Expected a {self.tp} but got {doc.__class__.__name__}")
+        return doc
+
+    def __repr__(self) -> str:
+        return str(self.tp)
+
+
+class _ArrayLoader(_Loader):
+    def __init__(self, items: _Loader) -> None:
+        self.items: Final = items
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> list[Any]:
+        if not isinstance(doc, MutableSequence):
+            raise ValidationException(
+                f"Value is a {convert_typing(extract_type(type(doc)))}, "
+                f"but valid type for this field is an array."
+            )
+        r: Final[list[Any]] = []
+        errors: Final[list[SchemaSaladException]] = []
+        fields: Final[list[str]] = []
+        for i in range(0, len(doc)):
+            try:
+                lf = _load_field(
+                    doc[i], _UnionLoader([self, self.items]), baseuri, loadingOptions, lc=lc
+                )
+                flatten = loadingOptions.container != "@list"
+                if flatten and isinstance(lf, MutableSequence):
+                    r.extend(lf)
+                else:
+                    r.append(lf)
+
+                if isinstance(doc[i], CommentedMap):
+                    if doc[i].get("id") is not None:
+                        if doc[i].get("id") in fields:
+                            errors.append(
+                                ValidationException(
+                                    f"Duplicate field {doc[i].get('id')!r}",
+                                    SourceLine(doc[i], "id", str),
+                                    [],
+                                )
+                            )
+                        else:
+                            fields.append(doc[i].get("id"))
+
+            except ValidationException as e:
+                e = ValidationException(
+                    "array item is invalid because", SourceLine(doc, i, str), [e]
+                )
+                errors.append(e)
+        if errors:
+            raise ValidationException("", None, errors)
+        return r
+
+    def __repr__(self) -> str:
+        return f"array<{self.items}>"
+
+
+class _MapLoader(_Loader):
+    def __init__(
+        self,
+        values: _Loader,
+        name: str | None = None,
+        container: str | None = None,
+        no_link_check: bool | None = None,
+    ) -> None:
+        self.values: Final = values
+        self.name: Final = name
+        self.container: Final = container
+        self.no_link_check: Final = no_link_check
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> dict[str, Any]:
+        if not isinstance(doc, MutableMapping):
+            raise ValidationException(f"Expected a map, was {type(doc)}")
+        if self.container is not None or self.no_link_check is not None:
+            loadingOptions = LoadingOptions(
+                copyfrom=loadingOptions, container=self.container, no_link_check=self.no_link_check
+            )
+        r: Final[dict[str, Any]] = {}
+        errors: Final[list[SchemaSaladException]] = []
+        for k, v in doc.items():
+            try:
+                lf = _load_field(v, self.values, baseuri, loadingOptions, lc)
+                r[k] = lf
+            except ValidationException as e:
+                errors.append(e.with_sourceline(SourceLine(doc, k, str)))
+        if errors:
+            raise ValidationException("", None, errors)
+        return r
+
+    def __repr__(self) -> str:
+        return self.name if self.name is not None else f"map<string, {self.values}>"
+
+
+class _EnumLoader(_Loader):
+    def __init__(self, symbols: Sequence[str], name: str) -> None:
+        self.symbols: Final = symbols
+        self.name: Final = name
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> str:
+        if doc in self.symbols:
+            return cast(str, doc)
+        raise ValidationException(f"Expected one of {self.symbols}")
+
+    def __repr__(self) -> str:
+        return self.name
+
+
+class _SecondaryDSLLoader(_Loader):
+    def __init__(self, inner: _Loader) -> None:
+        self.inner: Final = inner
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        r: Final[list[dict[str, Any]]] = []
+        match doc:
+            case MutableSequence() as dlist:
+                for d in dlist:
+                    if isinstance(d, str):
+                        if d.endswith("?"):
+                            r.append({"pattern": d[:-1], "required": False})
+                        else:
+                            r.append({"pattern": d})
+                    elif isinstance(d, dict):
+                        new_dict1: dict[str, Any] = {}
+                        dict_copy = copy.deepcopy(d)
+                        if "pattern" in dict_copy:
+                            new_dict1["pattern"] = dict_copy.pop("pattern")
+                        else:
+                            raise ValidationException(
+                                f"Missing pattern in secondaryFiles specification entry: {d}"
+                            )
+                        new_dict1["required"] = (
+                            dict_copy.pop("required") if "required" in dict_copy else None
+                        )
+
+                        if len(dict_copy):
+                            raise ValidationException(
+                                "Unallowed values in secondaryFiles specification entry: {}".format(
+                                    dict_copy
+                                )
+                            )
+                        r.append(new_dict1)
+
+                    else:
+                        raise ValidationException(
+                            "Expected a string or sequence of (strings or mappings)."
+                        )
+            case MutableMapping() as decl:
+                new_dict2 = {}
+                doc_copy = copy.deepcopy(decl)
+                if "pattern" in doc_copy:
+                    new_dict2["pattern"] = doc_copy.pop("pattern")
+                else:
+                    raise ValidationException(
+                        f"Missing pattern in secondaryFiles specification entry: {decl}"
+                    )
+                new_dict2["required"] = doc_copy.pop("required") if "required" in doc_copy else None
+
+                if len(doc_copy):
+                    raise ValidationException(
+                        f"Unallowed values in secondaryFiles specification entry: {doc_copy}"
+                    )
+                r.append(new_dict2)
+
+            case str(decl):
+                if decl.endswith("?"):
+                    r.append({"pattern": decl[:-1], "required": False})
+                else:
+                    r.append({"pattern": decl})
+            case _:
+                raise ValidationException("Expected str or sequence of str")
+        return self.inner.load(r, baseuri, loadingOptions, docRoot, lc=lc)
+
+
+class _RecordLoader(_Loader, Generic[S]):
+    def __init__(
+        self,
+        classtype: type[S],
+        container: str | None = None,
+        no_link_check: bool | None = None,
+    ) -> None:
+        self.classtype: Final = classtype
+        self.container: Final = container
+        self.no_link_check: Final = no_link_check
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> S:
+        if not isinstance(doc, MutableMapping):
+            raise ValidationException(
+                f"Value is a {convert_typing(extract_type(type(doc)))}, "
+                f"but valid type for this field is an object."
+            )
+        if self.container is not None or self.no_link_check is not None:
+            loadingOptions = LoadingOptions(
+                copyfrom=loadingOptions, container=self.container, no_link_check=self.no_link_check
+            )
+        return self.classtype.fromDoc(doc, baseuri, loadingOptions, docRoot=docRoot)
+
+    def __repr__(self) -> str:
+        return str(self.classtype.__name__)
+
+
+class _ExpressionLoader(_Loader):
+    def __init__(self, items: type[str]) -> None:
+        self.items: Final = items
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> str:
+        if not isinstance(doc, str):
+            raise ValidationException(
+                f"Value is a {convert_typing(extract_type(type(doc)))}, "
+                f"but valid type for this field is a str."
+            )
+        else:
+            return doc
+
+
+class _UnionLoader(_Loader):
+    def __init__(self, alternates: Sequence[_Loader], name: str | None = None) -> None:
+        self.alternates = alternates
+        self.name: Final = name
+
+    def add_loaders(self, loaders: Sequence[_Loader]) -> None:
+        self.alternates = tuple(loader for loader in chain(self.alternates, loaders))
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        errors: Final = []
+
+        if lc is None:
+            lc = []
+
+        for t in self.alternates:
+            try:
+                return t.load(doc, baseuri, loadingOptions, docRoot=docRoot, lc=lc)
+            except ValidationException as e:
+                if isinstance(t, _ArrayLoader) and len(self.alternates) > 1:
+                    continue
+                if isinstance(doc, (CommentedMap, dict)):
+                    if "class" in doc:
+                        if str(doc.get("class")) == str(t):
+                            errors.append(
+                                ValidationException(
+                                    f"Object `{baseuri.split('/')[-1]}` is not valid because:",
+                                    SourceLine(doc, next(iter(doc)), str),
+                                    [e],
+                                )
+                            )
+                    else:
+                        if "array" in str(t):
+                            continue
+                        else:
+                            if "id" in doc:
+                                id = baseuri.split("/")[-1] + "#" + str(doc.get("id"))
+                                if "id" in lc:
+                                    errors.append(
+                                        ValidationException(
+                                            f"checking object `{id}` using `{t}`",
+                                            SourceLine(lc, "id", str),
+                                            [e],
+                                        )
+                                    )
+                                else:
+                                    errors.append(
+                                        ValidationException(
+                                            f"checking object `{id}` using `{t}`",
+                                            SourceLine(lc, doc.get("id"), str),
+                                            [e],
+                                        )
+                                    )
+                            else:
+                                if not isinstance(
+                                    t, (_PrimitiveLoader)
+                                ):  # avoids 'tried <class "NoneType"> was {x}' errors
+                                    errors.append(
+                                        ValidationException(f"tried `{t}` but", None, [e])
+                                    )
+                else:
+                    # avoids "tried <class "CWLType"> but x" and instead returns the values for parsing
+                    errors.append(ValidationException("", None, [e]))
+
+        if isinstance(doc, (CommentedMap, dict)) and "class" in doc:
+            if str(doc.get("class")) not in str(self.alternates):
+                errors.append(
+                    ValidationException(
+                        "Field `class` contains undefined reference to "
+                        + "`"
+                        + "/".join(baseuri.split("/")[0:-1])
+                        + "/"
+                        + str(doc.get("class"))
+                        + "`",
+                        SourceLine(doc, "class", str),
+                        [],
+                    )
+                )
+        raise ValidationException("", None, errors, "*")
+
+    def __repr__(self) -> str:
+        return self.name if self.name is not None else " | ".join(str(a) for a in self.alternates)
+
+
+class _URILoader(_Loader):
+    def __init__(
+        self,
+        inner: _Loader,
+        scoped_id: bool,
+        vocab_term: bool,
+        scoped_ref: int | None,
+        no_link_check: bool | None,
+    ) -> None:
+        self.inner: Final = inner
+        self.scoped_id: Final = scoped_id
+        self.vocab_term: Final = vocab_term
+        self.scoped_ref: Final = scoped_ref
+        self.no_link_check: Final = no_link_check
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        if self.no_link_check is not None:
+            loadingOptions = LoadingOptions(
+                copyfrom=loadingOptions, no_link_check=self.no_link_check
+            )
+        match doc:
+            case MutableSequence() as decl:
+                newdoc: Final = []
+                for i in decl:
+                    if isinstance(i, str):
+                        newdoc.append(
+                            _expand_url(
+                                i,
+                                baseuri,
+                                loadingOptions,
+                                self.scoped_id,
+                                self.vocab_term,
+                                self.scoped_ref,
+                            )
+                        )
+                    else:
+                        newdoc.append(i)
+                doc = newdoc
+            case str(decl):
+                doc = _expand_url(
+                    decl,
+                    baseuri,
+                    loadingOptions,
+                    self.scoped_id,
+                    self.vocab_term,
+                    self.scoped_ref,
+                )
+        if isinstance(doc, str):
+            if not loadingOptions.no_link_check:
+                errors: Final = []
+                try:
+                    if not loadingOptions.fetcher.check_exists(doc):
+                        errors.append(
+                            ValidationException(f"contains undefined reference to `{doc}`")
+                        )
+                except ValidationException:
+                    pass
+                if len(errors) > 0:
+                    raise ValidationException("", None, errors)
+        return self.inner.load(doc, baseuri, loadingOptions, lc=lc)
+
+
+class _TypeDSLLoader(_Loader):
+    def __init__(
+        self,
+        inner: _Loader,
+        refScope: int | None,
+        salad_version: str,
+    ) -> None:
+        self.inner: Final = inner
+        self.refScope: Final = refScope
+        self.salad_version: Final = salad_version
+
+    def resolve(
+        self,
+        doc: str,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+    ) -> list[dict[str, Any] | str] | dict[str, Any] | str:
+        doc_ = doc
+        optional = False
+        if doc_.endswith("?"):
+            optional = True
+            doc_ = doc_[0:-1]
+
+        if doc_.endswith("[]"):
+            salad_versions: Final = [int(v) for v in self.salad_version[1:].split(".")]
+            items: list[dict[str, Any] | str] | dict[str, Any] | str = ""
+            rest: Final = doc_[0:-2]
+            if salad_versions < [1, 3]:
+                if rest.endswith("[]"):
+                    # To show the error message with the original type
+                    return doc
+                else:
+                    items = _expand_url(
+                        rest,
+                        baseuri,
+                        loadingOptions,
+                        False,
+                        True,
+                        self.refScope,
+                    )
+            else:
+                items = self.resolve(rest, baseuri, loadingOptions)
+                if isinstance(items, str):
+                    items = _expand_url(
+                        items,
+                        baseuri,
+                        loadingOptions,
+                        False,
+                        True,
+                        self.refScope,
+                    )
+            expanded: dict[str, Any] | str = {"type": "array", "items": items}
+        else:
+            expanded = _expand_url(
+                doc_,
+                baseuri,
+                loadingOptions,
+                False,
+                True,
+                self.refScope,
+            )
+
+        if optional:
+            return ["null", expanded]
+        else:
+            return expanded
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        if isinstance(doc, MutableSequence):
+            r: Final[list[Any]] = []
+            for d in doc:
+                if isinstance(d, str):
+                    resolved = self.resolve(d, baseuri, loadingOptions)
+                    if isinstance(resolved, MutableSequence):
+                        for i in resolved:
+                            if i not in r:
+                                r.append(i)
+                    else:
+                        if resolved not in r:
+                            r.append(resolved)
+                else:
+                    r.append(d)
+            doc = r
+        elif isinstance(doc, str):
+            doc = self.resolve(doc, baseuri, loadingOptions)
+
+        return self.inner.load(doc, baseuri, loadingOptions, lc=lc)
+
+
+class _IdMapLoader(_Loader):
+    def __init__(self, inner: _Loader, mapSubject: str, mapPredicate: str | None) -> None:
+        self.inner: Final = inner
+        self.mapSubject: Final = mapSubject
+        self.mapPredicate: Final = mapPredicate
+
+    def load(
+        self,
+        doc: Any,
+        baseuri: str,
+        loadingOptions: LoadingOptions,
+        docRoot: str | None = None,
+        lc: Any | None = None,
+    ) -> Any:
+        if isinstance(doc, MutableMapping):
+            r: Final[list[Any]] = []
+            for k in doc.keys():
+                val = doc[k]
+                if isinstance(val, CommentedMap):
+                    v = copy.copy(val)
+                    v.lc.data = val.lc.data
+                    v.lc.filename = val.lc.filename
+                    v[self.mapSubject] = k
+                    r.append(v)
+                elif isinstance(val, MutableMapping):
+                    v2 = copy.copy(val)
+                    v2[self.mapSubject] = k
+                    r.append(v2)
+                else:
+                    if self.mapPredicate:
+                        v3 = {self.mapPredicate: val}
+                        v3[self.mapSubject] = k
+                        r.append(v3)
+                    else:
+                        raise ValidationException("No mapPredicate")
+            doc = r
+        return self.inner.load(doc, baseuri, loadingOptions, lc=lc)
+
+
+def _document_load(
+    loader: _Loader,
+    doc: str | MutableMapping[str, Any] | MutableSequence[Any],
+    baseuri: str,
+    loadingOptions: LoadingOptions,
+    addl_metadata_fields: MutableSequence[str] | None = None,
+) -> tuple[Any, LoadingOptions]:
+    if isinstance(doc, str):
+        return _document_load_by_url(
+            loader,
+            loadingOptions.fetcher.urljoin(baseuri, doc),
+            loadingOptions,
+            addl_metadata_fields=addl_metadata_fields,
+        )
+
+    if isinstance(doc, MutableMapping):
+        addl_metadata: Final = {}
+        if addl_metadata_fields is not None:
+            for mf in addl_metadata_fields:
+                if mf in doc:
+                    addl_metadata[mf] = doc[mf]
+
+        docuri: Final = baseuri
+        if "$base" in doc:
+            baseuri = doc["$base"]
+
+        loadingOptions = LoadingOptions(
+            copyfrom=loadingOptions,
+            namespaces=doc.get("$namespaces", None),
+            schemas=doc.get("$schemas", None),
+            baseuri=doc.get("$base", None),
+            addl_metadata=addl_metadata,
+        )
+
+        doc2: Final = copy.copy(doc)
+        if "$namespaces" in doc2:
+            doc2.pop("$namespaces")
+        if "$schemas" in doc2:
+            doc2.pop("$schemas")
+        if "$base" in doc2:
+            doc2.pop("$base")
+
+        if "$graph" in doc2:
+            loadingOptions.idx[baseuri] = (
+                loader.load(doc2["$graph"], baseuri, loadingOptions),
+                loadingOptions,
+            )
+        else:
+            loadingOptions.idx[baseuri] = (
+                loader.load(doc2, baseuri, loadingOptions, docRoot=baseuri),
+                loadingOptions,
+            )
+
+        if docuri != baseuri:
+            loadingOptions.idx[docuri] = loadingOptions.idx[baseuri]
+
+        return loadingOptions.idx[baseuri]
+
+    if isinstance(doc, MutableSequence):
+        loadingOptions.idx[baseuri] = (
+            loader.load(doc, baseuri, loadingOptions),
+            loadingOptions,
+        )
+        return loadingOptions.idx[baseuri]
+
+    raise ValidationException(
+        "Expected URI string, MutableMapping or MutableSequence, got %s" % type(doc)
+    )
+
+
+def _document_load_by_url(
+    loader: _Loader,
+    url: str,
+    loadingOptions: LoadingOptions,
+    addl_metadata_fields: MutableSequence[str] | None = None,
+) -> tuple[Any, LoadingOptions]:
+    if url in loadingOptions.idx:
+        return loadingOptions.idx[url]
+
+    doc_url, frg = urldefrag(url)
+
+    text: Final = loadingOptions.fetcher.fetch_text(doc_url)
+    textIO: Final = StringIO(text)
+    textIO.name = str(doc_url)
+    yaml: Final = yaml_no_ts()
+    result: Final = yaml.load(textIO)
+    add_lc_filename(result, doc_url)
+
+    loadingOptions = LoadingOptions(copyfrom=loadingOptions, fileuri=doc_url)
+
+    _document_load(
+        loader,
+        result,
+        doc_url,
+        loadingOptions,
+        addl_metadata_fields=addl_metadata_fields,
+    )
+
+    return loadingOptions.idx[url]
+
+
+def _expand_url(
+    url: str,
+    base_url: str,
+    loadingOptions: LoadingOptions,
+    scoped_id: bool = False,
+    vocab_term: bool = False,
+    scoped_ref: int | None = None,
+) -> str:
+    if url in ("@id", "@type"):
+        return url
+
+    vocab = _vocab | loadingOptions.vocab
+    if vocab_term and url in vocab:
+        return url
+
+    if bool(vocab) and ":" in url:
+        prefix: Final = url.split(":")[0]
+        if prefix in vocab:
+            url = vocab[prefix] + url[len(prefix) + 1 :]
+
+    split1: Final = urlsplit(url)
+
+    if (
+        (bool(split1.scheme) and split1.scheme in loadingOptions.fetcher.supported_schemes())
+        or url.startswith("$(")
+        or url.startswith("${")
+    ):
+        pass
+    elif scoped_id and not bool(split1.fragment):
+        splitbase1: Final = urlsplit(base_url)
+        frg: str
+        if bool(splitbase1.fragment):
+            frg = splitbase1.fragment + "/" + split1.path
+        else:
+            frg = split1.path
+        pt: Final = splitbase1.path if splitbase1.path != "" else "/"
+        url = urlunsplit((splitbase1.scheme, splitbase1.netloc, pt, splitbase1.query, frg))
+    elif scoped_ref is not None and not bool(split1.fragment):
+        splitbase2: Final = urlsplit(base_url)
+        sp = splitbase2.fragment.split("/")
+        n = scoped_ref
+        while n > 0 and len(sp) > 0:
+            sp.pop()
+            n -= 1
+        sp.append(url)
+        url = urlunsplit(
+            (
+                splitbase2.scheme,
+                splitbase2.netloc,
+                splitbase2.path,
+                splitbase2.query,
+                "/".join(sp),
+            )
+        )
+    else:
+        url = loadingOptions.fetcher.urljoin(base_url, url)
+
+    if vocab_term:
+        split2: Final = urlsplit(url)
+        if bool(split2.scheme):
+            if url in (rvocab := _rvocab | loadingOptions.rvocab):
+                return rvocab[url]
+        else:
+            raise ValidationException(f"Term {url!r} not in vocabulary")
+
+    return url
+
+
+def _load_field(
+    val: Any | None,
+    fieldtype: "_Loader",
+    baseuri: str,
+    loadingOptions: LoadingOptions,
+    lc: Any | None = None,
+) -> Any:
+    """Load field."""
+    if isinstance(val, MutableMapping):
+        if "$import" in val:
+            if loadingOptions.fileuri is None:
+                raise SchemaSaladException("Cannot load $import without fileuri")
+            url1: Final = loadingOptions.fetcher.urljoin(loadingOptions.fileuri, val["$import"])
+            result, metadata = _document_load_by_url(
+                fieldtype,
+                url1,
+                loadingOptions,
+            )
+            loadingOptions.imports.append(url1)
+            return result
+        if "$include" in val:
+            if loadingOptions.fileuri is None:
+                raise SchemaSaladException("Cannot load $import without fileuri")
+            url2: Final = loadingOptions.fetcher.urljoin(loadingOptions.fileuri, val["$include"])
+            val = loadingOptions.fetcher.fetch_text(url2)
+            loadingOptions.includes.append(url2)
+    return fieldtype.load(val, baseuri, loadingOptions, lc=lc)
 
 
 def parser_info() -> str:
@@ -119,7 +899,7 @@ class RecordField(Documented):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_strtype_True_False_None_None,
                     baseuri,
@@ -175,7 +955,7 @@ class RecordField(Documented):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -223,7 +1003,7 @@ class RecordField(Documented):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_2,
                 baseuri,
@@ -275,14 +1055,8 @@ class RecordField(Documented):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -385,7 +1159,7 @@ class RecordSchema(Saveable):
         fields = None
         if "fields" in _doc:
             try:
-                fields = load_field(
+                fields = _load_field(
                     _doc.get("fields"),
                     idmap_fields_union_of_None_type_or_array_of_RecordFieldLoader,
                     baseuri,
@@ -433,7 +1207,7 @@ class RecordSchema(Saveable):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Record_nameLoader_2,
                 baseuri,
@@ -485,14 +1259,8 @@ class RecordSchema(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -603,7 +1371,7 @@ class EnumSchema(Saveable):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_union_of_None_type_or_strtype_True_False_None_None,
                     baseuri,
@@ -660,7 +1428,7 @@ class EnumSchema(Saveable):
             if _doc.get("symbols") is None:
                 raise ValidationException("missing required field `symbols`", None, [])
 
-            symbols = load_field(
+            symbols = _load_field(
                 _doc.get("symbols"),
                 uri_array_of_strtype_True_False_None_None,
                 baseuri,
@@ -708,7 +1476,7 @@ class EnumSchema(Saveable):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Enum_nameLoader_2,
                 baseuri,
@@ -760,14 +1528,8 @@ class EnumSchema(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -870,7 +1632,7 @@ class ArraySchema(Saveable):
             if _doc.get("items") is None:
                 raise ValidationException("missing required field `items`", None, [])
 
-            items = load_field(
+            items = _load_field(
                 _doc.get("items"),
                 uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_False_True_2_None,
                 baseuri,
@@ -918,7 +1680,7 @@ class ArraySchema(Saveable):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Array_nameLoader_2,
                 baseuri,
@@ -970,14 +1732,8 @@ class ArraySchema(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -1075,7 +1831,7 @@ class MapSchema(Saveable):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Map_nameLoader_2,
                 baseuri,
@@ -1123,7 +1879,7 @@ class MapSchema(Saveable):
             if _doc.get("values") is None:
                 raise ValidationException("missing required field `values`", None, [])
 
-            values = load_field(
+            values = _load_field(
                 _doc.get("values"),
                 uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_False_True_2_None,
                 baseuri,
@@ -1175,14 +1931,8 @@ class MapSchema(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -1280,7 +2030,7 @@ class UnionSchema(Saveable):
             if _doc.get("names") is None:
                 raise ValidationException("missing required field `names`", None, [])
 
-            names = load_field(
+            names = _load_field(
                 _doc.get("names"),
                 uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_False_True_2_None,
                 baseuri,
@@ -1328,7 +2078,7 @@ class UnionSchema(Saveable):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Union_nameLoader_2,
                 baseuri,
@@ -1380,14 +2130,8 @@ class UnionSchema(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -1533,7 +2277,7 @@ class JsonldPredicate(Saveable):
         _id = None
         if "_id" in _doc:
             try:
-                _id = load_field(
+                _id = _load_field(
                     _doc.get("_id"),
                     uri_union_of_None_type_or_strtype_True_False_None_None,
                     baseuri,
@@ -1580,7 +2324,7 @@ class JsonldPredicate(Saveable):
         _type = None
         if "_type" in _doc:
             try:
-                _type = load_field(
+                _type = _load_field(
                     _doc.get("_type"),
                     union_of_None_type_or_strtype,
                     baseuri,
@@ -1627,7 +2371,7 @@ class JsonldPredicate(Saveable):
         _container = None
         if "_container" in _doc:
             try:
-                _container = load_field(
+                _container = _load_field(
                     _doc.get("_container"),
                     union_of_None_type_or_strtype,
                     baseuri,
@@ -1674,7 +2418,7 @@ class JsonldPredicate(Saveable):
         identity = None
         if "identity" in _doc:
             try:
-                identity = load_field(
+                identity = _load_field(
                     _doc.get("identity"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -1721,7 +2465,7 @@ class JsonldPredicate(Saveable):
         noLinkCheck = None
         if "noLinkCheck" in _doc:
             try:
-                noLinkCheck = load_field(
+                noLinkCheck = _load_field(
                     _doc.get("noLinkCheck"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -1768,7 +2512,7 @@ class JsonldPredicate(Saveable):
         mapSubject = None
         if "mapSubject" in _doc:
             try:
-                mapSubject = load_field(
+                mapSubject = _load_field(
                     _doc.get("mapSubject"),
                     union_of_None_type_or_strtype,
                     baseuri,
@@ -1815,7 +2559,7 @@ class JsonldPredicate(Saveable):
         mapPredicate = None
         if "mapPredicate" in _doc:
             try:
-                mapPredicate = load_field(
+                mapPredicate = _load_field(
                     _doc.get("mapPredicate"),
                     union_of_None_type_or_strtype,
                     baseuri,
@@ -1862,7 +2606,7 @@ class JsonldPredicate(Saveable):
         refScope = None
         if "refScope" in _doc:
             try:
-                refScope = load_field(
+                refScope = _load_field(
                     _doc.get("refScope"),
                     union_of_None_type_or_inttype,
                     baseuri,
@@ -1909,7 +2653,7 @@ class JsonldPredicate(Saveable):
         typeDSL = None
         if "typeDSL" in _doc:
             try:
-                typeDSL = load_field(
+                typeDSL = _load_field(
                     _doc.get("typeDSL"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -1956,7 +2700,7 @@ class JsonldPredicate(Saveable):
         secondaryFilesDSL = None
         if "secondaryFilesDSL" in _doc:
             try:
-                secondaryFilesDSL = load_field(
+                secondaryFilesDSL = _load_field(
                     _doc.get("secondaryFilesDSL"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -2003,7 +2747,7 @@ class JsonldPredicate(Saveable):
         subscope = None
         if "subscope" in _doc:
             try:
-                subscope = load_field(
+                subscope = _load_field(
                     _doc.get("subscope"),
                     union_of_None_type_or_strtype,
                     baseuri,
@@ -2055,14 +2799,8 @@ class JsonldPredicate(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -2237,7 +2975,7 @@ class SpecializeDef(Saveable):
             if _doc.get("specializeFrom") is None:
                 raise ValidationException("missing required field `specializeFrom`", None, [])
 
-            specializeFrom = load_field(
+            specializeFrom = _load_field(
                 _doc.get("specializeFrom"),
                 uri_strtype_False_False_1_None,
                 baseuri,
@@ -2285,7 +3023,7 @@ class SpecializeDef(Saveable):
             if _doc.get("specializeTo") is None:
                 raise ValidationException("missing required field `specializeTo`", None, [])
 
-            specializeTo = load_field(
+            specializeTo = _load_field(
                 _doc.get("specializeTo"),
                 uri_strtype_False_False_1_None,
                 baseuri,
@@ -2337,14 +3075,8 @@ class SpecializeDef(Saveable):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -2483,7 +3215,7 @@ class SaladRecordField(RecordField):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_strtype_True_False_None_None,
                     baseuri,
@@ -2539,7 +3271,7 @@ class SaladRecordField(RecordField):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -2587,7 +3319,7 @@ class SaladRecordField(RecordField):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_2,
                 baseuri,
@@ -2634,7 +3366,7 @@ class SaladRecordField(RecordField):
         jsonldPredicate = None
         if "jsonldPredicate" in _doc:
             try:
-                jsonldPredicate = load_field(
+                jsonldPredicate = _load_field(
                     _doc.get("jsonldPredicate"),
                     union_of_None_type_or_strtype_or_JsonldPredicateLoader,
                     baseuri,
@@ -2681,7 +3413,7 @@ class SaladRecordField(RecordField):
         default = None
         if "default" in _doc:
             try:
-                default = load_field(
+                default = _load_field(
                     _doc.get("default"),
                     union_of_None_type_or_Any_type,
                     baseuri,
@@ -2733,14 +3465,8 @@ class SaladRecordField(RecordField):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -2912,7 +3638,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_strtype_True_False_None_None,
                     baseuri,
@@ -2968,7 +3694,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         inVocab = None
         if "inVocab" in _doc:
             try:
-                inVocab = load_field(
+                inVocab = _load_field(
                     _doc.get("inVocab"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -3015,7 +3741,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         fields = None
         if "fields" in _doc:
             try:
-                fields = load_field(
+                fields = _load_field(
                     _doc.get("fields"),
                     idmap_fields_union_of_None_type_or_array_of_SaladRecordFieldLoader,
                     baseuri,
@@ -3063,7 +3789,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Record_nameLoader_2,
                 baseuri,
@@ -3110,7 +3836,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -3157,7 +3883,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         docParent = None
         if "docParent" in _doc:
             try:
-                docParent = load_field(
+                docParent = _load_field(
                     _doc.get("docParent"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -3204,7 +3930,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         docChild = None
         if "docChild" in _doc:
             try:
-                docChild = load_field(
+                docChild = _load_field(
                     _doc.get("docChild"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_None_None,
                     baseuri,
@@ -3251,7 +3977,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         docAfter = None
         if "docAfter" in _doc:
             try:
-                docAfter = load_field(
+                docAfter = _load_field(
                     _doc.get("docAfter"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -3298,7 +4024,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         jsonldPredicate = None
         if "jsonldPredicate" in _doc:
             try:
-                jsonldPredicate = load_field(
+                jsonldPredicate = _load_field(
                     _doc.get("jsonldPredicate"),
                     union_of_None_type_or_strtype_or_JsonldPredicateLoader,
                     baseuri,
@@ -3345,7 +4071,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         documentRoot = None
         if "documentRoot" in _doc:
             try:
-                documentRoot = load_field(
+                documentRoot = _load_field(
                     _doc.get("documentRoot"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -3392,7 +4118,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         abstract = None
         if "abstract" in _doc:
             try:
-                abstract = load_field(
+                abstract = _load_field(
                     _doc.get("abstract"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -3439,7 +4165,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         extends = None
         if "extends" in _doc:
             try:
-                extends = load_field(
+                extends = _load_field(
                     _doc.get("extends"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_1_None,
                     baseuri,
@@ -3486,7 +4212,7 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
         specialize = None
         if "specialize" in _doc:
             try:
-                specialize = load_field(
+                specialize = _load_field(
                     _doc.get("specialize"),
                     idmap_specialize_union_of_None_type_or_array_of_SpecializeDefLoader,
                     baseuri,
@@ -3538,14 +4264,8 @@ class SaladRecordSchema(NamedType, RecordSchema, SchemaDefinedType):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -3773,7 +4493,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_union_of_None_type_or_strtype_True_False_None_None,
                     baseuri,
@@ -3829,7 +4549,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         inVocab = None
         if "inVocab" in _doc:
             try:
-                inVocab = load_field(
+                inVocab = _load_field(
                     _doc.get("inVocab"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -3877,7 +4597,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
             if _doc.get("symbols") is None:
                 raise ValidationException("missing required field `symbols`", None, [])
 
-            symbols = load_field(
+            symbols = _load_field(
                 _doc.get("symbols"),
                 uri_array_of_strtype_True_False_None_None,
                 baseuri,
@@ -3925,7 +4645,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Enum_nameLoader_2,
                 baseuri,
@@ -3972,7 +4692,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -4019,7 +4739,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         docParent = None
         if "docParent" in _doc:
             try:
-                docParent = load_field(
+                docParent = _load_field(
                     _doc.get("docParent"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -4066,7 +4786,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         docChild = None
         if "docChild" in _doc:
             try:
-                docChild = load_field(
+                docChild = _load_field(
                     _doc.get("docChild"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_None_None,
                     baseuri,
@@ -4113,7 +4833,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         docAfter = None
         if "docAfter" in _doc:
             try:
-                docAfter = load_field(
+                docAfter = _load_field(
                     _doc.get("docAfter"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -4160,7 +4880,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         jsonldPredicate = None
         if "jsonldPredicate" in _doc:
             try:
-                jsonldPredicate = load_field(
+                jsonldPredicate = _load_field(
                     _doc.get("jsonldPredicate"),
                     union_of_None_type_or_strtype_or_JsonldPredicateLoader,
                     baseuri,
@@ -4207,7 +4927,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         documentRoot = None
         if "documentRoot" in _doc:
             try:
-                documentRoot = load_field(
+                documentRoot = _load_field(
                     _doc.get("documentRoot"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -4254,7 +4974,7 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
         extends = None
         if "extends" in _doc:
             try:
-                extends = load_field(
+                extends = _load_field(
                     _doc.get("extends"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_1_None,
                     baseuri,
@@ -4306,14 +5026,8 @@ class SaladEnumSchema(NamedType, EnumSchema, SchemaDefinedType):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -4518,7 +5232,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_strtype_True_False_None_None,
                     baseuri,
@@ -4574,7 +5288,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         inVocab = None
         if "inVocab" in _doc:
             try:
-                inVocab = load_field(
+                inVocab = _load_field(
                     _doc.get("inVocab"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -4622,7 +5336,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Map_nameLoader_2,
                 baseuri,
@@ -4670,7 +5384,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
             if _doc.get("values") is None:
                 raise ValidationException("missing required field `values`", None, [])
 
-            values = load_field(
+            values = _load_field(
                 _doc.get("values"),
                 uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_False_True_2_None,
                 baseuri,
@@ -4717,7 +5431,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -4764,7 +5478,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         docParent = None
         if "docParent" in _doc:
             try:
-                docParent = load_field(
+                docParent = _load_field(
                     _doc.get("docParent"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -4811,7 +5525,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         docChild = None
         if "docChild" in _doc:
             try:
-                docChild = load_field(
+                docChild = _load_field(
                     _doc.get("docChild"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_None_None,
                     baseuri,
@@ -4858,7 +5572,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         docAfter = None
         if "docAfter" in _doc:
             try:
-                docAfter = load_field(
+                docAfter = _load_field(
                     _doc.get("docAfter"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -4905,7 +5619,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         jsonldPredicate = None
         if "jsonldPredicate" in _doc:
             try:
-                jsonldPredicate = load_field(
+                jsonldPredicate = _load_field(
                     _doc.get("jsonldPredicate"),
                     union_of_None_type_or_strtype_or_JsonldPredicateLoader,
                     baseuri,
@@ -4952,7 +5666,7 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
         documentRoot = None
         if "documentRoot" in _doc:
             try:
-                documentRoot = load_field(
+                documentRoot = _load_field(
                     _doc.get("documentRoot"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -5004,14 +5718,8 @@ class SaladMapSchema(NamedType, MapSchema, SchemaDefinedType):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -5207,7 +5915,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_strtype_True_False_None_None,
                     baseuri,
@@ -5263,7 +5971,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         inVocab = None
         if "inVocab" in _doc:
             try:
-                inVocab = load_field(
+                inVocab = _load_field(
                     _doc.get("inVocab"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -5311,7 +6019,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
             if _doc.get("names") is None:
                 raise ValidationException("missing required field `names`", None, [])
 
-            names = load_field(
+            names = _load_field(
                 _doc.get("names"),
                 uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_False_True_2_None,
                 baseuri,
@@ -5359,7 +6067,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Union_nameLoader_2,
                 baseuri,
@@ -5406,7 +6114,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -5453,7 +6161,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         docParent = None
         if "docParent" in _doc:
             try:
-                docParent = load_field(
+                docParent = _load_field(
                     _doc.get("docParent"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -5500,7 +6208,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         docChild = None
         if "docChild" in _doc:
             try:
-                docChild = load_field(
+                docChild = _load_field(
                     _doc.get("docChild"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_None_None,
                     baseuri,
@@ -5547,7 +6255,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         docAfter = None
         if "docAfter" in _doc:
             try:
-                docAfter = load_field(
+                docAfter = _load_field(
                     _doc.get("docAfter"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -5594,7 +6302,7 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
         documentRoot = None
         if "documentRoot" in _doc:
             try:
-                documentRoot = load_field(
+                documentRoot = _load_field(
                     _doc.get("documentRoot"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -5646,14 +6354,8 @@ class SaladUnionSchema(NamedType, UnionSchema, DocType):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -5832,7 +6534,7 @@ class Documentation(NamedType, DocType):
         name = None
         if "name" in _doc:
             try:
-                name = load_field(
+                name = _load_field(
                     _doc.get("name"),
                     uri_strtype_True_False_None_None,
                     baseuri,
@@ -5888,7 +6590,7 @@ class Documentation(NamedType, DocType):
         inVocab = None
         if "inVocab" in _doc:
             try:
-                inVocab = load_field(
+                inVocab = _load_field(
                     _doc.get("inVocab"),
                     union_of_None_type_or_booltype,
                     baseuri,
@@ -5935,7 +6637,7 @@ class Documentation(NamedType, DocType):
         doc = None
         if "doc" in _doc:
             try:
-                doc = load_field(
+                doc = _load_field(
                     _doc.get("doc"),
                     union_of_None_type_or_strtype_or_array_of_strtype,
                     baseuri,
@@ -5982,7 +6684,7 @@ class Documentation(NamedType, DocType):
         docParent = None
         if "docParent" in _doc:
             try:
-                docParent = load_field(
+                docParent = _load_field(
                     _doc.get("docParent"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -6029,7 +6731,7 @@ class Documentation(NamedType, DocType):
         docChild = None
         if "docChild" in _doc:
             try:
-                docChild = load_field(
+                docChild = _load_field(
                     _doc.get("docChild"),
                     uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_None_None,
                     baseuri,
@@ -6076,7 +6778,7 @@ class Documentation(NamedType, DocType):
         docAfter = None
         if "docAfter" in _doc:
             try:
-                docAfter = load_field(
+                docAfter = _load_field(
                     _doc.get("docAfter"),
                     uri_union_of_None_type_or_strtype_False_False_None_None,
                     baseuri,
@@ -6124,7 +6826,7 @@ class Documentation(NamedType, DocType):
             if _doc.get("type") is None:
                 raise ValidationException("missing required field `type`", None, [])
 
-            type_ = load_field(
+            type_ = _load_field(
                 _doc.get("type"),
                 typedsl_Documentation_nameLoader_2,
                 baseuri,
@@ -6176,14 +6878,8 @@ class Documentation(NamedType, DocType):
                         ValidationException("mapping with implicit null key")
                     )
                 elif ":" in k:
-                    ex = expand_url(
-                        k,
-                        "",
-                        loadingOptions,
-                        _vocab,
-                        _rvocab,
-                        scoped_id=False,
-                        vocab_term=False,
+                    ex = _expand_url(
+                        k, "", loadingOptions, scoped_id=False, vocab_term=False
                     )
                     extension_fields[ex] = _doc[k]
                 else:
@@ -6395,9 +7091,7 @@ union_of_None_type_or_strtype_or_array_of_strtype: Final = _UnionLoader(
         array_of_strtype,
     )
 )
-uri_strtype_True_False_None_None: Final = _URILoader(
-    strtype, True, False, None, None, _vocab, _rvocab
-)
+uri_strtype_True_False_None_None: Final = _URILoader(strtype, True, False, None, None)
 union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype: (
     Final
 ) = _UnionLoader(
@@ -6436,8 +7130,6 @@ typedsl_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_o
     union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype,
     2,
     "v1.1",
-    _vocab,
-    _rvocab,
 )
 array_of_RecordFieldLoader: Final = _ArrayLoader(RecordFieldLoader)
 union_of_None_type_or_array_of_RecordFieldLoader: Final = _UnionLoader(
@@ -6450,9 +7142,7 @@ idmap_fields_union_of_None_type_or_array_of_RecordFieldLoader: Final = _IdMapLoa
     union_of_None_type_or_array_of_RecordFieldLoader, "name", "type"
 )
 Record_nameLoader: Final = _EnumLoader(("record",), "Record_name")
-typedsl_Record_nameLoader_2: Final = _TypeDSLLoader(
-    Record_nameLoader, 2, "v1.1", _vocab, _rvocab
-)
+typedsl_Record_nameLoader_2: Final = _TypeDSLLoader(Record_nameLoader, 2, "v1.1")
 union_of_None_type_or_strtype: Final = _UnionLoader(
     (
         None_type,
@@ -6460,15 +7150,13 @@ union_of_None_type_or_strtype: Final = _UnionLoader(
     )
 )
 uri_union_of_None_type_or_strtype_True_False_None_None: Final = _URILoader(
-    union_of_None_type_or_strtype, True, False, None, None, _vocab, _rvocab
+    union_of_None_type_or_strtype, True, False, None, None
 )
 uri_array_of_strtype_True_False_None_None: Final = _URILoader(
-    array_of_strtype, True, False, None, None, _vocab, _rvocab
+    array_of_strtype, True, False, None, None
 )
 Enum_nameLoader: Final = _EnumLoader(("enum",), "Enum_name")
-typedsl_Enum_nameLoader_2: Final = _TypeDSLLoader(
-    Enum_nameLoader, 2, "v1.1", _vocab, _rvocab
-)
+typedsl_Enum_nameLoader_2: Final = _TypeDSLLoader(Enum_nameLoader, 2, "v1.1")
 uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_or_array_of_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_ArraySchemaLoader_or_MapSchemaLoader_or_UnionSchemaLoader_or_strtype_False_True_2_None: (
     Final
 ) = _URILoader(
@@ -6477,21 +7165,13 @@ uri_union_of_PrimitiveTypeLoader_or_RecordSchemaLoader_or_EnumSchemaLoader_or_Ar
     True,
     2,
     None,
-    _vocab,
-    _rvocab,
 )
 Array_nameLoader: Final = _EnumLoader(("array",), "Array_name")
-typedsl_Array_nameLoader_2: Final = _TypeDSLLoader(
-    Array_nameLoader, 2, "v1.1", _vocab, _rvocab
-)
+typedsl_Array_nameLoader_2: Final = _TypeDSLLoader(Array_nameLoader, 2, "v1.1")
 Map_nameLoader: Final = _EnumLoader(("map",), "Map_name")
-typedsl_Map_nameLoader_2: Final = _TypeDSLLoader(
-    Map_nameLoader, 2, "v1.1", _vocab, _rvocab
-)
+typedsl_Map_nameLoader_2: Final = _TypeDSLLoader(Map_nameLoader, 2, "v1.1")
 Union_nameLoader: Final = _EnumLoader(("union",), "Union_name")
-typedsl_Union_nameLoader_2: Final = _TypeDSLLoader(
-    Union_nameLoader, 2, "v1.1", _vocab, _rvocab
-)
+typedsl_Union_nameLoader_2: Final = _TypeDSLLoader(Union_nameLoader, 2, "v1.1")
 union_of_None_type_or_booltype: Final = _UnionLoader(
     (
         None_type,
@@ -6504,21 +7184,13 @@ union_of_None_type_or_inttype: Final = _UnionLoader(
         inttype,
     )
 )
-uri_strtype_False_False_1_None: Final = _URILoader(
-    strtype, False, False, 1, None, _vocab, _rvocab
-)
+uri_strtype_False_False_1_None: Final = _URILoader(strtype, False, False, 1, None)
 uri_union_of_None_type_or_strtype_False_False_None_None: Final = _URILoader(
-    union_of_None_type_or_strtype, False, False, None, None, _vocab, _rvocab
+    union_of_None_type_or_strtype, False, False, None, None
 )
 uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_None_None: Final = (
     _URILoader(
-        union_of_None_type_or_strtype_or_array_of_strtype,
-        False,
-        False,
-        None,
-        None,
-        _vocab,
-        _rvocab,
+        union_of_None_type_or_strtype_or_array_of_strtype, False, False, None, None
     )
 )
 union_of_None_type_or_strtype_or_JsonldPredicateLoader: Final = _UnionLoader(
@@ -6545,15 +7217,7 @@ idmap_fields_union_of_None_type_or_array_of_SaladRecordFieldLoader: Final = (
     _IdMapLoader(union_of_None_type_or_array_of_SaladRecordFieldLoader, "name", "type")
 )
 uri_union_of_None_type_or_strtype_or_array_of_strtype_False_False_1_None: Final = (
-    _URILoader(
-        union_of_None_type_or_strtype_or_array_of_strtype,
-        False,
-        False,
-        1,
-        None,
-        _vocab,
-        _rvocab,
-    )
+    _URILoader(union_of_None_type_or_strtype_or_array_of_strtype, False, False, 1, None)
 )
 array_of_SpecializeDefLoader: Final = _ArrayLoader(SpecializeDefLoader)
 union_of_None_type_or_array_of_SpecializeDefLoader: Final = _UnionLoader(
@@ -6571,7 +7235,7 @@ idmap_specialize_union_of_None_type_or_array_of_SpecializeDefLoader: Final = (
 )
 Documentation_nameLoader: Final = _EnumLoader(("documentation",), "Documentation_name")
 typedsl_Documentation_nameLoader_2: Final = _TypeDSLLoader(
-    Documentation_nameLoader, 2, "v1.1", _vocab, _rvocab
+    Documentation_nameLoader, 2, "v1.1"
 )
 union_of_SaladRecordSchemaLoader_or_SaladEnumSchemaLoader_or_SaladMapSchemaLoader_or_SaladUnionSchemaLoader_or_DocumentationLoader: (
     Final
